deleteのbuttonを追加してconfirmダイアログを表示する機能を実装 文字カウンターで8割でorange、10割でredに変化する機能を実装

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+
 
 class PostController extends Controller
 {
@@ -30,6 +32,10 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        if(!auth()->check()){
+            abort(403,'ログインして投稿して下さい。');
+        }
+
         $validated = $request->validate([
             'title'=>'required|max:20',
             'body'=>'required|max:400',
@@ -63,11 +69,15 @@ class PostController extends Controller
      */
     public function update(Request $request,Post $post)
     {
+        if ($post->user_id !== auth()->id()) {
+            abort(403, 'この投稿を更新する権限がありません。');
+        }
+
         $validated = $request->validate([
             'title'=>'required|max:20',
             'body'=>'required|max:400',
         ]);
-        $validated['user_id']= auth()->id();
+        // $validated['user_id']= auth()->id();
 
         $post->update($validated);
 
@@ -80,6 +90,10 @@ class PostController extends Controller
      */
     public function destroy(Request $request,Post $post)
     {
+        if ($post->user_id !== auth()->id()) {
+            abort(403, 'この投稿を削除する権限がありません。');
+        }
+
         $post->delete();
         $request->session()->flash('message','削除しました');
         return redirect()->route('post.index');
